쿠폰 기능 수정중

+++ backend/TeamGoott/src/main/java/com/team/goott/owner/coupon/controller/OwnerCouponController.java
@@ -85,11 +85,10 @@ public class OwnerCouponController {
 	 // 쿠폰 삭제
     @DeleteMapping("/{couponId}") // 쿠폰 ID를 경로 변수로 받음
     public ResponseEntity<Object> deleteCoupon(HttpSession session, @PathVariable int couponId) {
-//    	log.info(couponId+"");
-//        StoreDTO storeSession = (StoreDTO) session.getAttribute("store");
-//        if (storeSession == null) {
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다."); // 로그인 필요
-//        }
+        StoreDTO storeSession = (StoreDTO) session.getAttribute("store");
+        if (storeSession == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인이 필요합니다."); // 로그인 필요
+        }
         
         try {
             int result = ownerCouponService.deleteCoupon(couponId); // 삭제 메서드 호출
